SJ-4145 move update rest address after registration/reregistration
