+++ core/sj-mesos-framework/src/main/scala/com/bwsw/sj/mesos/framework/task/status/SuccessHandler.scala
@@ -18,20 +18,16 @@
  */
 package com.bwsw.sj.mesos.framework.task.status
 
-import com.bwsw.sj.common.dal.repository.ConnectionRepository
+import com.bwsw.sj.mesos.framework.mesos_api.{MasterState, Slave, SlaveState}
 import com.bwsw.sj.mesos.framework.schedule.FrameworkUtil
 import com.bwsw.sj.mesos.framework.task.TasksList
 import org.apache.mesos.Protos.{SlaveID, TaskID, TaskStatus}
 import com.bwsw.sj.mesos.framework.task.StatusHandler
-import com.bwsw.sj.mesos.framework.task.status.states.{MasterState, Slave, SlaveState}
-import scaldi.Injectable._
 
 object SuccessHandler {
 
   def process(status: TaskStatus): Unit = {
     val currentSlave = getCurrentSlave(status.getSlaveId)
-
-    updateInstanceRestAddress(currentSlave)
 
     val dir = extractSandbox(currentSlave, status.getTaskId)
     val dirUrl = s"http://${FrameworkUtil.master.get.getHostname}:${FrameworkUtil.master.get.getPort}/#/slaves/${currentSlave.id}/browse?path=$dir"
@@ -55,12 +51,5 @@ object SuccessHandler {
     val obj = StatusHandler.serializer.deserialize[SlaveState](slaveResponse)
     val framework = obj.frameworks.filter(framework => FrameworkUtil.frameworkId.get.contains(framework.id)).head
     framework.executors.filter(executor => executor.id == taskId.getValue).head.directory
-  }
-
-  def updateInstanceRestAddress(slave: Slave): Unit = {
-    val slaveHostname = slave.hostname
-    val address = s"http://$slaveHostname:${FrameworkUtil.instancePort.get}"
-    FrameworkUtil.instance.get.restAddress = Option(address)
-    inject[ConnectionRepository].getInstanceRepository.save(FrameworkUtil.instance.get.to)
   }
 }
